// src/app/models/product.model.ts
export interface Product {
  id: number;
  productName: string;
  description: string;
  category: string;
  brand: string;
  expiredDate: string;
  manufacturedDate: string;
  createdDate: string;
  quantity: number;
  unitPrice: number;
  // manufactureName: string;
  // daysForDeliver: number;
}